support local CI and add docstrings

+++ dev/testbranch.sh
@@ -1,28 +1,39 @@
 #!/bin/bash
 
 ## This script makes a copy of the Symata source and runs the test suite.
-## This script must be run from the ./dev directory
+## This script must be run from the ./dev directory.
+## To use this script, you must have "./" first in LOAD_PATH,
+## so that Julia finds the copy of the Symata package first.
 
-## Note: to use this script, you must have "./" first in LOAD_PATH
-## so that Julia finds the copy of the Symata package first
-
+## Get the first 9 characters of the HEAD commit
 cd ../ && COMMIT_HASH=$(git rev-parse --short=9 HEAD) && cd ./dev/
 
+copycode=false
+
+## Make a commit-unique test directory
 TEST_PACKAGE_DIR=./testsrc/Symata-${COMMIT_HASH}
+if [ ! -d $TEST_PACKAGE_DIR/ ]; then
+    mkdir $TEST_PACKAGE_DIR/
+    copycode=true
+fi
 
-echo mkdir $TEST_PACKAGE_DIR/
-mkdir $TEST_PACKAGE_DIR/
 
-echo rm -rf $TEST_PACKAGE_DIR/*
-rm -rf $TEST_PACKAGE_DIR/*
+if [ "$copycode" = true ]; then
+    ## Clean the test directory if it exists and add a package directory, Symata
+    rm -rf $TEST_PACKAGE_DIR/*
+    mkdir $TEST_PACKAGE_DIR/Symata/
+    ## Copy the source files to the test package directory
+    ## Project.toml must be present (15 Jul 2018) for loading modules correctly.
+    cp -a ../src ../sjtest ../pysrc ../symsrc ../Project.toml $TEST_PACKAGE_DIR/Symata/
+fi
 
-mkdir $TEST_PACKAGE_DIR/Symata/
+cd $TEST_PACKAGE_DIR/
 
-# # Project.toml must be present (15 Jul 2018) for loading modules correctly.
-cp -a ../src ../sjtest ../pysrc ../symsrc ../Project.toml $TEST_PACKAGE_DIR/Symata/
+# Important to not use "-i" here.
+julia -e '@time using Symata; @sym(Tests())' > >(tee -a stdout.log) 2> >(tee -a stderr.log >&2)
 
-# cd dev/testsrc && pwd
+## This buffers all output till the end.
+#julia -e '@time using Symata; @sym(Tests())' &> tests.log
 
-#cd testsrc && julia -e "@time using Symata; @time(@sym(Tests()))"
 
-exit 1
+# exit 1
